rename(`argon.js`): to match other files
